Separating complexityForPartialParallel from set_relation.cc and puting it into its own file
